Update index.js
omg... just ignore the last file. I have no idea how to use github apparently but the requested change I made should now be deleted an this is the better version to look over. It's still probably retarded but I missed a closing parentheses at the end of the "if" line.